Most Solr tests working.

+++ proxy/jaxb/src/main/java/org/ivdnt/blacklab/proxy/representation/CaptureGroup.java
@@ -7,9 +7,9 @@ import jakarta.xml.bind.annotation.XmlAccessorType;
 public class CaptureGroup {
     public String name;
 
-    public int start;
+    public long start;
 
-    public int end;
+    public long end;
 
     @Override
     public String toString() {
