+++ proxy/jaxb/src/main/java/org/ivdnt/blacklab/proxy/representation/MatchInfo.java
@@ -3,17 +3,17 @@ package org.ivdnt.blacklab.proxy.representation;
 import java.util.List;
 import java.util.Map;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 public class MatchInfo {
 
-    public int start;
+    public long start;
 
-    public int end;
+    public long end;
 
     /** What match info type this is (span, tag, relation or list) */
     @JsonInclude(JsonInclude.Include.NON_NULL)
